+++ torch_geometric_signed_directed/data/directed/__init__.py
@@ -1,2 +1,3 @@
 from .DSBM import DSBM
 from .DirectedData import DirectedData
+from .WebKB import load_WebKB
